+++ packages/adapter-node/index.d.ts
@@ -1,12 +1,17 @@
-type BuildOptions = import('esbuild').BuildOptions;
-declare function plugin(options?: {
+import { Adapter } from '@sveltejs/kit';
+import { BuildOptions } from 'esbuild';
+
+interface AdapterOptions {
+	entryPoint?: string;
 	out?: string;
 	precompress?: boolean;
 	env?: {
+		path?: string;
 		host?: string;
 		port?: string;
 	};
-	esbuild?: (defaultOptions: BuildOptions) => Promise<BuildOptions> | BuildOptions;
-}): import('@sveltejs/kit').Adapter;
+	esbuild?: (options: BuildOptions) => Promise<BuildOptions> | BuildOptions;
+}
 
+declare function plugin(options?: AdapterOptions): Adapter;
 export = plugin;