+++ packages/kit/src/packaging/index.js
@@ -1,9 +1,10 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import globrex from 'globrex';
 import { createRequire } from 'module';
+
 import { preprocess } from 'svelte/compiler';
-import { mkdirp, rimraf } from '../utils/filesystem.js';
+
+import { mkdirp, rimraf, walk } from '../utils/filesystem.js';
 
 const essential_files = ['README', 'LICENSE', 'CHANGELOG', '.gitignore', '.npmignore'];
 
@@ -12,41 +13,49 @@ const essential_files = ['README', 'LICENSE', 'CHANGELOG', '.gitignore', '.npmig
  * @param {string} cwd
  */
 export async function make_package(config, cwd = process.cwd()) {
-	rimraf(path.join(cwd, config.kit.package.dir));
+	const abs_package_dir = path.join(cwd, config.kit.package.dir);
+	rimraf(abs_package_dir);
 
 	if (config.kit.package.emitTypes) {
 		// Generate type definitions first so hand-written types can overwrite generated ones
 		await emit_dts(config);
+		// Resolve aliases, TS leaves them as-is
+		const files = walk(abs_package_dir);
+		for (const file of files) {
+			const filename = path.join(abs_package_dir, file);
+			const source = fs.readFileSync(filename, 'utf8');
+			fs.writeFileSync(filename, resolve_$lib_alias(file, source, config));
+		}
 	}
-
-	const files_filter = create_filter(config.kit.package.files);
-	const exports_filter = create_filter({
-		...config.kit.package.exports,
-		exclude: [...config.kit.package.exports.exclude, '*.d.ts']
-	});
 
 	const files = walk(config.kit.files.lib);
 
 	const pkg = JSON.parse(fs.readFileSync(path.join(cwd, 'package.json'), 'utf8'));
 
 	delete pkg.scripts;
-	pkg.type = 'module'; // type must be 'module'
+	pkg.type = 'module';
 
-	const user_defined_exports = 'exports' in pkg;
+	/** @type {Record<string, string>} */
+	const generated = { './package.json': './package.json' };
 
-	// We still want to always predefine some exports
-	// like package.json that is used by other packages
-	if (!pkg.exports) pkg.exports = {};
-	pkg.exports['./package.json'] = './package.json';
+	/** @type {Record<string, string>} */
+	const clashes = {};
+	let contains_svelte_files = false;
 
 	for (const file of files) {
-		if (!files_filter(file)) continue;
+		const ext = path.extname(file);
+		const normalized = file.replace(/\\/g, '/');
+		const svelte_ext = config.extensions.find((ext) => file.endsWith(ext)); // unlike `ext`, could be e.g. `.svelte.md`
+
+		if (!config.kit.package.files(normalized)) {
+			const dts_file = (svelte_ext ? file : file.slice(0, -ext.length)) + '.d.ts';
+			const dts_path = path.join(abs_package_dir, dts_file);
+			if (fs.existsSync(dts_path)) fs.unlinkSync(dts_path);
+			continue;
+		}
 
 		const filename = path.join(config.kit.files.lib, file);
 		const source = fs.readFileSync(filename, 'utf8');
-
-		const ext = path.extname(file);
-		const svelte_ext = config.extensions.find((ext) => file.endsWith(ext)); // unlike `ext`, could be e.g. `.svelte.md`
 
 		/** @type {string} */
 		let out_file;
@@ -58,13 +67,14 @@ export async function make_package(config, cwd = process.cwd()) {
 			// it's a Svelte component
 			out_file = file.slice(0, -svelte_ext.length) + '.svelte';
 			out_contents = config.preprocess
-				? (await preprocess(source, config.preprocess, { filename })).code
+				? strip_lang_tags((await preprocess(source, config.preprocess, { filename })).code)
 				: source;
+			contains_svelte_files = true;
 		} else if (ext === '.ts' && file.endsWith('.d.ts')) {
 			// TypeScript's declaration emit won't copy over the d.ts files, so we do it here
 			out_file = file;
 			out_contents = source;
-			if (fs.existsSync(path.join(cwd, config.kit.package.dir, out_file))) {
+			if (fs.existsSync(path.join(abs_package_dir, out_file))) {
 				console.warn(
 					'Found already existing file from d.ts generation for ' +
 						out_file +
@@ -78,23 +88,56 @@ export async function make_package(config, cwd = process.cwd()) {
 			out_file = file;
 			out_contents = source;
 		}
+		out_contents = resolve_$lib_alias(out_file, out_contents, config);
 
-		write(path.join(cwd, config.kit.package.dir, out_file), out_contents);
+		write(path.join(abs_package_dir, out_file), out_contents);
 
-		if (!user_defined_exports && exports_filter(file)) {
+		if (config.kit.package.exports(normalized)) {
+			const original = `$lib/${normalized}`;
 			const entry = `./${out_file.replace(/\\/g, '/')}`;
-			const key = entry.endsWith('/index.js') ? entry.slice(0, -'/index.js'.length) : entry;
-			pkg.exports[key] = entry;
+			const key = entry.replace(/\/index\.js$|(\/[^/]+)\.js$/, '$1');
+
+			if (clashes[key]) {
+				throw new Error(
+					`Duplicate "${key}" export. Please remove or rename either ${clashes[key]} or ${original}`
+				);
+			}
+
+			generated[key] = entry;
+			clashes[key] = original;
 		}
 	}
 
-	const main = pkg.exports['./index.js'] || pkg.exports['./index.svelte'];
+	pkg.exports = { ...generated, ...pkg.exports };
 
-	if (!user_defined_exports && main) {
-		pkg.exports['.'] = main;
+	if (!pkg.svelte && contains_svelte_files) {
+		// Several heuristics in Kit/vite-plugin-svelte to tell Vite to mark Svelte packages
+		// rely on the "svelte" property. Vite/Rollup/Webpack plugin can all deal with it.
+		// See https://github.com/sveltejs/kit/issues/1959 for more info and related threads.
+		if (pkg.exports['.']) {
+			const svelte_export =
+				typeof pkg.exports['.'] === 'string'
+					? pkg.exports['.']
+					: pkg.exports['.'].import || pkg.exports['.'].default;
+			if (svelte_export) {
+				pkg.svelte = svelte_export;
+			} else {
+				console.warn(
+					'Cannot generate a "svelte" entry point because ' +
+						'the "." entry in "exports" is not a string. ' +
+						'If you set it by hand, please also set one of the options as a "svelte" entry point'
+				);
+			}
+		} else {
+			console.warn(
+				'Cannot generate a "svelte" entry point because ' +
+					'the "." entry in "exports" is missing. ' +
+					'Please specify one or set a "svelte" entry point yourself'
+			);
+		}
 	}
 
-	write(path.join(cwd, config.kit.package.dir, 'package.json'), JSON.stringify(pkg, null, '  '));
+	write(path.join(abs_package_dir, 'package.json'), JSON.stringify(pkg, null, '  '));
 
 	const whitelist = fs.readdirSync(cwd).filter((file) => {
 		const lowercased = file.toLowerCase();
@@ -104,8 +147,73 @@ export async function make_package(config, cwd = process.cwd()) {
 		const full_path = path.join(cwd, pathname);
 		if (fs.lstatSync(full_path).isDirectory()) continue; // just to be sure
 
-		const package_path = path.join(cwd, config.kit.package.dir, pathname);
+		const package_path = path.join(abs_package_dir, pathname);
 		if (!fs.existsSync(package_path)) fs.copyFileSync(full_path, package_path);
+	}
+}
+
+/**
+ * Resolves the `$lib` alias.
+ *
+ * TODO: make this more generic to also handle other aliases the user could have defined
+ * via `kit.vite.resolve.alias`. Also investage how to do this in a more robust way
+ * (right now regex string replacement is used).
+ * For more discussion see https://github.com/sveltejs/kit/pull/2453
+ *
+ * @param {string} file Relative to the lib root
+ * @param {string} content
+ * @param {import('types/config').ValidatedConfig} config
+ * @returns {string}
+ */
+function resolve_$lib_alias(file, content, config) {
+	/**
+	 * @param {string} match
+	 * @param {string} _
+	 * @param {string} import_path
+	 */
+	const replace_import_path = (match, _, import_path) => {
+		if (!import_path.startsWith('$lib/')) {
+			return match;
+		}
+
+		const full_path = path.join(config.kit.files.lib, file);
+		const full_import_path = path.join(config.kit.files.lib, import_path.slice('$lib/'.length));
+		let resolved = path.relative(path.dirname(full_path), full_import_path).replace(/\\/g, '/');
+		resolved = resolved.startsWith('.') ? resolved : './' + resolved;
+		return match.replace(import_path, resolved);
+	};
+	content = content.replace(/from\s+('|")([^"';,]+?)\1/g, replace_import_path);
+	content = content.replace(/import\s*\(\s*('|")([^"';,]+?)\1\s*\)/g, replace_import_path);
+	return content;
+}
+
+/**
+ * Strip out lang="X" or type="text/X" tags. Doing it here is only a temporary solution.
+ * See https://github.com/sveltejs/kit/issues/2450 for ideas for places where it's handled better.
+ *
+ * @param {string} content
+ */
+function strip_lang_tags(content) {
+	strip_lang_tag('script');
+	strip_lang_tag('style');
+	return content;
+
+	/**
+	 * @param {string} tagname
+	 */
+	function strip_lang_tag(tagname) {
+		const regexp = new RegExp(
+			`/<!--[^]*?-->|<${tagname}(\\s[^]*?)?(?:>([^]*?)<\\/${tagname}>|\\/>)`,
+			'g'
+		);
+		content = content.replace(regexp, (tag, attributes) => {
+			const idx = tag.indexOf(attributes);
+			return (
+				tag.substring(0, idx) +
+				attributes.replace(/\s(type|lang)=(["']).*?\2/, ' ') +
+				tag.substring(idx + attributes.length)
+			);
+		});
 	}
 }
 
@@ -163,49 +271,6 @@ function load_tsconfig(filename, ts) {
 }
 
 /**
- * @param {{
- *   include: string[];
- *   exclude: string[];
- * }} options
- */
-function create_filter(options) {
-	const include = options.include.map((str) => str && globrex(str));
-	const exclude = options.exclude.map((str) => str && globrex(str));
-
-	/** @param {string} str */
-	const filter = (str) =>
-		include.some((glob) => glob && glob.regex.test(str)) &&
-		!exclude.some((glob) => glob && glob.regex.test(str));
-
-	return filter;
-}
-
-/** @param {string} cwd */
-function walk(cwd) {
-	/** @type {string[]} */
-	const all_files = [];
-
-	/** @param {string} dir */
-	function walk_dir(dir) {
-		const files = fs.readdirSync(path.join(cwd, dir));
-
-		for (const file of files) {
-			const joined = path.join(dir, file);
-			const stats = fs.statSync(path.join(cwd, joined));
-
-			if (stats.isDirectory()) {
-				walk_dir(joined);
-			} else {
-				all_files.push(joined);
-			}
-		}
-	}
-
-	walk_dir('');
-	return all_files;
-}
-
-/**
  * @param {string} file
  * @param {string} contents
  */
@@ -220,7 +285,7 @@ function write(file, contents) {
 export async function emit_dts(config) {
 	const require = createRequire(import.meta.url);
 	const emit = await try_load_svelte2tsx();
-	emit({
+	await emit({
 		libRoot: config.kit.files.lib,
 		svelteShimsPath: require.resolve('svelte2tsx/svelte-shims.d.ts'),
 		declarationDir: config.kit.package.dir