+++ packages/kit/src/runtime/server/page/load_node.js
@@ -12,7 +12,7 @@ const s = JSON.stringify;
  *   page: import('types/page').Page;
  *   node: import('types/internal').SSRNode;
  *   $session: any;
- *   context: Record<string, any>;
+ *   stuff: Record<string, any>;
  *   prerender_enabled: boolean;
  *   is_leaf: boolean;
  *   is_error: boolean;
@@ -29,7 +29,7 @@ export async function load_node({
 	page,
 	node,
 	$session,
-	context,
+	stuff,
 	prerender_enabled,
 	is_leaf,
 	is_error,
@@ -48,6 +48,11 @@ export async function load_node({
 	 * }>}
 	 */
 	const fetched = [];
+
+	/**
+	 * @type {string[]}
+	 */
+	let set_cookie_headers = [];
 
 	let loaded;
 
@@ -120,7 +125,9 @@ export async function load_node({
 				} else if (resolved.startsWith('/') && !resolved.startsWith('//')) {
 					const relative = resolved;
 
-					const headers = /** @type {import('types/helper').Headers} */ ({ ...opts.headers });
+					const headers = /** @type {import('types/helper').RequestHeaders} */ ({
+						...opts.headers
+					});
 
 					// TODO: fix type https://github.com/node-fetch/node-fetch/issues/1113
 					if (opts.credentials !== 'omit') {
@@ -149,7 +156,7 @@ export async function load_node({
 							method: opts.method || 'GET',
 							headers,
 							path: relative,
-							rawBody: new TextEncoder().encode(/** @type {string} */ (opts.body)),
+							rawBody: opts.body == null ? null : new TextEncoder().encode(opts.body),
 							query: new URLSearchParams(search)
 						},
 						options,
@@ -164,9 +171,12 @@ export async function load_node({
 							state.prerender.dependencies.set(relative, rendered);
 						}
 
+						// Set-Cookie must be filtered out (done below) and that's the only header value that
+						// can be an array so we know we have only simple values
+						// https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie
 						response = new Response(rendered.body, {
 							status: rendered.status,
-							headers: rendered.headers
+							headers: /** @type {Record<string, string>} */ (rendered.headers)
 						});
 					}
 				} else {
@@ -211,10 +221,14 @@ export async function load_node({
 							async function text() {
 								const body = await response.text();
 
-								/** @type {import('types/helper').Headers} */
+								/** @type {import('types/helper').ResponseHeaders} */
 								const headers = {};
 								for (const [key, value] of response.headers) {
-									if (key !== 'etag' && key !== 'set-cookie') headers[key] = value;
+									if (key === 'set-cookie') {
+										set_cookie_headers = set_cookie_headers.concat(value);
+									} else if (key !== 'etag') {
+										headers[key] = value;
+									}
 								}
 
 								if (!opts.body || typeof opts.body === 'string') {
@@ -255,7 +269,7 @@ export async function load_node({
 					})
 				);
 			},
-			context: { ...context }
+			stuff: { ...stuff }
 		};
 
 		if (is_error) {
@@ -279,8 +293,9 @@ export async function load_node({
 	return {
 		node,
 		loaded: normalize(loaded),
-		context: loaded.context || context,
+		stuff: loaded.stuff || stuff,
 		fetched,
+		set_cookie_headers,
 		uses_credentials
 	};
 }