+++ packages/kit/src/runtime/server/page/render.js
@@ -1,5 +1,6 @@
 import devalue from 'devalue';
 import { writable } from 'svelte/store';
+import { coalesce_to_error } from '../../../utils/error.js';
 import { hash } from '../../hash.js';
 
 const s = JSON.stringify;
@@ -172,10 +173,10 @@ export async function render_response({
 
 					return `<script ${attributes}>${json}</script>`;
 				})
-				.join('\n\n\t\t\t')}
-		`.replace(/^\t{2}/gm, '');
+				.join('\n\n\t')}
+		`;
 
-	/** @type {import('types/helper').Headers} */
+	/** @type {import('types/helper').ResponseHeaders} */
 	const headers = {
 		'content-type': 'text/html'
 	};
@@ -203,7 +204,7 @@ function try_serialize(data, fail) {
 	try {
 		return devalue(data);
 	} catch (err) {
-		if (fail) fail(err);
+		if (fail) fail(coalesce_to_error(err));
 		return null;
 	}
 }