+++ packages/adapter-static/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sveltejs/adapter-static
 
+## 1.0.0-next.20
+
+### Patch Changes
+
+- [chore] upgrade to Svelte 3.43.0" ([#2474](https://github.com/sveltejs/kit/pull/2474))
+
+## 1.0.0-next.19
+
+### Patch Changes
+
+- [chore] add links to repository and homepage to package.json ([#2425](https://github.com/sveltejs/kit/pull/2425))
+
+## 1.0.0-next.18
+
+### Patch Changes
+
+- [chore] export package.json from adapters ([#2351](https://github.com/sveltejs/kit/pull/2351))
+
+## 1.0.0-next.17
+
+### Patch Changes
+
+- Clear output before adapting ([#2260](https://github.com/sveltejs/kit/pull/2260))
+
 ## 1.0.0-next.16
 
 ### Patch Changes